change cta name, refactoring layout, improving tests
